==> notebooks/1_preprocessing.ipynb <==
import pandas as pd

# **Caesarian data classification**

Загрузка таблицы

cl_df = pd.read_csv('../data/caesarian/caesarian.csv', sep=',')
cl_df

Получение информации

cl_df.info()

cl_df.isnull().sum()

*Предобработка*

cl_df = pd.get_dummies(cl_df,prefix=['Delivery_number'], columns=['Delivery_number'])
cl_df = pd.get_dummies(cl_df,prefix=['Blood_of_Pressure'], columns=['Blood_of_Pressure'])
cl_df = pd.get_dummies(cl_df,prefix=['Delivery_time'], columns=['Delivery_time'])
cl_df

Проверка оставшихся столбцов на бинарность

cl_df['Heart_Problem'].unique()

cl_df['Caesarian'].unique()

Сохранение в csv

cl_df.to_csv('../data/caesarian/caesarian_preprocessed.csv')

# ***Toxicity fish regression***

Загрузка таблицы

re_df = pd.read_csv('../data/qsar/qsar_fish_toxicity.csv', sep=';')
re_df

Информация о таблице

re_df.info()

re_df.isnull().sum()

Проверка столбцов типа int на бинарность

re_df['NdsCH'].unique()

re_df['NdssC'].unique()

Предобработка

re_df = pd.get_dummies(re_df,prefix=['NdsCH'], columns=['NdsCH'])
re_df = pd.get_dummies(re_df,prefix=['NdssC'], columns=['NdssC'])
re_df

Сохранение в csv

re_df.to_csv('../data/qsar/qsar_fish_toxicity_preprocessed.csv')
