==> notebooks/3_regression.ipynb <==
# Предобработка данных

import numpy as np
import pandas as pd
from sklearn import metrics
from sklearn.model_selection import GridSearchCV

df = pd.DataFrame
df = pd.read_csv('../qsar/qsar_fish_toxicity_preprocessed.csv' )
del df['Unnamed: 0']
df

df.info()

X = df.iloc[:, :4].join(df.iloc[:, 5:]).values
y = df.iloc[:, 4].values
print(X)

print(y)

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.30)

X_train

from sklearn.preprocessing import StandardScaler
scaler = StandardScaler()
scaler.fit(X_train)

X_train = scaler.transform(X_train)
X_test = scaler.transform(X_test)
X_train


# Regression

## Линейная регрессия

[Ссылка на документацию](https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LinearRegression.html)

from sklearn.linear_model import LinearRegression
lreg = LinearRegression().fit(X_train, y_train)
y_pred = lreg.predict(X_test)

model_score = lreg.score(X_test,y_test)
r_square = metrics.r2_score(y_test, y_pred)
mse = metrics.mean_squared_error(y_test, y_pred)
mae = metrics.mean_absolute_error(y_test, y_pred)


print("Model Score            : ", model_score)
print("R Square               : ", r_square)
print("Mean Squared Error     : ", mse)
print("Root Mean Squared Error: ", mse**(1/2))
print("Mean Absolute Error    : ", mae)


## Регрессия дерева решений

[Ссылка на документацию](https://scikit-learn.org/stable/modules/generated/sklearn.tree.DecisionTreeRegressor.html)

from sklearn.tree import DecisionTreeRegressor

clf = GridSearchCV(DecisionTreeRegressor(), {'max_depth': range(2,10,1)})

clf.fit(X_train, y_train)

clf.best_params_

dtr = DecisionTreeRegressor(max_depth=4).fit(X_train, y_train)
y_pred = dtr.predict(X_test)


model_score = dtr.score(X_test,y_test)
r_square = metrics.r2_score(y_test, y_pred)
mse = metrics.mean_squared_error(y_test, y_pred)
mae = metrics.mean_absolute_error(y_test, y_pred)

print("Model Score            : ", model_score)
print("R Square               : ", r_square)
print("Mean Squared Error     : ", mse)
print("Root Mean Squared Error: ", mse**(1/2))
print("Mean Absolute Error    : ", mae)


## LASSO 
[Сслыка на документацию](https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.Lasso.html)

from sklearn.linear_model import Lasso
ls = Lasso().fit(X_train, y_train)
y_pred = ls.predict(X_test)

model_score = ls.score(X_test,y_test)
r_square = metrics.r2_score(y_test, y_pred)
mse = metrics.mean_squared_error(y_test, y_pred)
mae = metrics.mean_absolute_error(y_test, y_pred)


print("Model Score            : ", model_score)
print("R Square               : ", r_square)
print("Mean Squared Error     : ", mse)
print("Root Mean Squared Error: ", mse**(1/2))
print("Mean Absolute Error    : ", mae)

## Гребневая регрессия
[Ссылка на документацию](https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.Ridge.html)

from sklearn.linear_model import Ridge
Rid = Ridge().fit(X_train, y_train)
y_pred = Rid.predict(X_test)

model_score = Rid.score(X_test,y_test)
r_square = metrics.r2_score(y_test, y_pred)
mse = metrics.mean_squared_error(y_test, y_pred)
mae = metrics.mean_absolute_error(y_test, y_pred)


print("Model Score            : ", model_score)
print("R Square               : ", r_square)
print("Mean Squared Error     : ", mse)
print("Root Mean Squared Error: ", mse**(1/2))
print("Mean Absolute Error    : ", mae)

## Elastic Net регрессия
[Ссылка на документацию](https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.ElasticNet.html)

from sklearn.linear_model import ElasticNet
Rln = ElasticNet().fit(X_train, y_train)
y_pred = Rln.predict(X_test)

model_score = Rln.score(X_test,y_test)
r_square = metrics.r2_score(y_test, y_pred)
mse = metrics.mean_squared_error(y_test, y_pred)
mae = metrics.mean_absolute_error(y_test, y_pred)


print("Model Score            : ", model_score)
print("R Square               : ", r_square)
print("Mean Squared Error     : ", mse)
print("Root Mean Squared Error: ", mse**(1/2))
print("Mean Absolute Error    : ", mae)

# Вывод

Линейная и гребневая регрессия имеют наибольшую точность из представленных. Однако все представленные методы имеют досточно низкую точность из-за разброса данных
