==> notebooks/4_ensemble_classification.ipynb <==
# Подготовка данных

import pandas as pd
import numpy as np
from sklearn import metrics
from sklearn import ensemble

df = pd.DataFrame
df = pd.read_csv('../data/caesarian/caesarian_preprocessed.csv' )
del df['Unnamed: 0']
df.head()

df['Caesarian'].unique()

df1 = df[:]
del df1['Caesarian']

X = df1.iloc[:, :].values
y = df.iloc[:, 2].values
print(X)

print(y)

from sklearn.model_selection import train_test_split
from collections import Counter

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20, stratify=y)

print(Counter(y_train).values())
print(Counter(y_test).values())


from sklearn.preprocessing import StandardScaler
scaler = StandardScaler()
scaler.fit(X_train)

X_train = scaler.transform(X_train)
X_test = scaler.transform(X_test)
X_train

#Композиции алгоритмов

from sklearn.model_selection import GridSearchCV

##Бэггинг

from sklearn.ensemble import BaggingClassifier
from sklearn.naive_bayes import GaussianNB

%%time
clf = GaussianNB()
bgc = BaggingClassifier(clf)

bagging_parameters = {'n_estimators' : range(10, 50, 10), 'max_features' : range(1,5,1)}
bagging = GridSearchCV(bgc, bagging_parameters)
bagging.fit(X_train, y_train)

bagging.best_params_

y_pred = bagging.predict(X_test)
print(metrics.confusion_matrix(y_test, y_pred))
print(metrics.classification_report(y_test, y_pred))

##Бустинг

from sklearn.ensemble import GradientBoostingClassifier

%%time
gbc = GradientBoostingClassifier()
boosting_parameters = {'n_estimators' : range(5,20,5), 'max_depth' : range(1,8,1)}
boosting = GridSearchCV(gbc, boosting_parameters)
boosting.fit(X_train, y_train)

boosting.best_params_

y_pred = boosting.predict(X_test)
print(metrics.confusion_matrix(y_test, y_pred))
print(metrics.classification_report(y_test, y_pred))

##Стекинг

from sklearn.ensemble import StackingClassifier
from sklearn.neighbors import KNeighborsClassifier

knc_parameters = {'n_neighbors': range(1, 10, 1)}

knc = KNeighborsClassifier()
clf_knc = GridSearchCV(knc, knc_parameters)

clf_knc.fit(X_train, y_train)
clf_knc.best_params_

%%time
estimators = [
('gnb', GaussianNB()),
('knn', KNeighborsClassifier(n_neighbors = 2))]

sc = StackingClassifier(estimators=estimators)
sc.fit(X_train, y_train)
y_pred = sc.predict(X_test)
print(metrics.confusion_matrix(y_test, y_pred))
print(metrics.classification_report(y_test, y_pred))

#Вывод

Алгоритм бустинг показал самую высокую точность.

Алгоритм бэггинг сработал немного хуже.

Стекинг показал самую низкую точность.

В целом, точность алгоритмов довольно низкая из-за небольшой подборки данных.
