==> notebooks/4_ensemble_regression.ipynb <==
#Подготовка данных

import pandas as pd
import numpy as np
from sklearn import metrics
from sklearn import ensemble

df = pd.DataFrame
df = pd.read_csv('../data/qsar/qsar_fish_toxicity_preprocessed.csv' )
del df['Unnamed: 0']
df

df.info()

X = df.iloc[:, :3].join(df.iloc[:, 4]).values
y = df.iloc[:, 3].values
print(X)

print(y)

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.30)

X_train

from sklearn.preprocessing import StandardScaler
scaler = StandardScaler()
scaler.fit(X_train)

X_train = scaler.transform(X_train)
X_test = scaler.transform(X_test)
X_train

#Композиции алгоритмов

from sklearn.model_selection import GridSearchCV

##Бэггинг

from sklearn.ensemble import BaggingRegressor
from sklearn.tree import DecisionTreeRegressor

clf = GridSearchCV(DecisionTreeRegressor(), {'max_depth': range(1,20,2)})
clf.fit(X_train, y_train)
clf.best_params_

%%time
dtr = DecisionTreeRegressor(max_depth=5)
bgc = BaggingRegressor(dtr)

bagging_parameters = {'n_estimators' : range(10, 50, 10), 'max_features' : range(1,8,1)}
bagging = GridSearchCV(bgc, bagging_parameters)
bagging.fit(X_train, y_train)

bagging.best_params_

y_pred = bagging.predict(X_test)

model_score = bagging.score(X_test,y_test)
mse = metrics.mean_squared_error(y_test, y_pred)
mae = metrics.mean_absolute_error(y_test, y_pred)

print("Model Score            : ", model_score)
print("Mean Squared Error     : ", mse)
print("Root Mean Squared Error: ", mse**(1/2))
print("Mean Absolute Error    : ", mae)

##Бустинг

from sklearn.ensemble import GradientBoostingRegressor

%%time
gbc = GradientBoostingRegressor()
boosting_parameters = {'n_estimators' : range(5,20,5), 'max_depth' : range(1,8,1)}
boosting = GridSearchCV(gbc, boosting_parameters)
boosting.fit(X_train, y_train)

boosting.best_params_

y_pred = boosting.predict(X_test)

model_score = boosting.score(X_test,y_test)
mse = metrics.mean_squared_error(y_test, y_pred)
mae = metrics.mean_absolute_error(y_test, y_pred)

print("Model Score            : ", model_score)
print("Mean Squared Error     : ", mse)
print("Root Mean Squared Error: ", mse**(1/2))
print("Mean Absolute Error    : ", mae)

##Стекинг

from sklearn.ensemble import StackingRegressor
from sklearn.linear_model import LinearRegression
from sklearn.linear_model import Ridge

%%time
estimators = [('lreg', LinearRegression()),('Rid', Ridge())]
sr = StackingRegressor(estimators=estimators)
stacking = sr.fit(X_train, y_train)

y_pred = boosting.predict(X_test)

model_score = stacking.score(X_test,y_test)
mse = metrics.mean_squared_error(y_test, y_pred)
mae = metrics.mean_absolute_error(y_test, y_pred)

print("Model Score            : ", model_score)
print("Mean Squared Error     : ", mse)
print("Root Mean Squared Error: ", mse**(1/2))
print("Mean Absolute Error    : ", mae)

#Вывод

Все алгоритмы показали приблизительно одинаковые результаты, однако, наилучшую точность показал бустинг
